requirements.txt
requests>=2.31.0
langgraph>=0.0.10
langchain-core>=0.1.0
psutil>=5.9.0
python-dotenv>=1.0.0
ruff>=0.1.0
